requirements.txt
jsonschema>=4.22.0
openai>=1.40.0
requests>=2.31.0
typer>=0.9.0
sseclient-py>=1.8.0
